Feat: crud admins finalizado

+++ app/Http/Controllers/Admin/AdminController.php
@@ -10,24 +10,28 @@ use Illuminate\Http\Request;
 
 class AdminController extends Controller
 {
-    protected AdminServiceInterface $service;
+    private AdminServiceInterface $service;
+    private array $searchFields = [
+        'nome',
+        'email',
+    ];
 
     public function __construct(AdminServiceInterface $service)
     {
         $this->service = $service;
     }
 
-    public function index(){
-        $admins = $this->service->all();
-        return view('admin.painel.admins.admins-list')->with('admins', $admins);
+    public function index(Request $request){
+        $admins = $this->service->allFilter($request->busca, $this->searchFields, 1);
+        return view('admin.painel.admins.admins-list')->with(['admins' => $admins, 'request' => $request->all()]);
     }
 
     public function show(Admin $admin){
-        return view('admin.painel.admins.show-admins')->with('admin', $admin);
+        return view('admin.painel.admins.admins-show')->with('admin', $admin);
     }
 
     public function create(){
-        return view('admin.painel.admins.create-admins');
+        return view('admin.painel.admins.admins-create');
     }
 
     public function store(AdminRequest $request){
@@ -36,7 +40,7 @@ class AdminController extends Controller
     }
 
     public function edit(Admin $admin){
-        return view('admin.painel.admins.edit-admins')->with('admin', $admin);
+        return view('admin.painel.admins.admins-edit')->with('admin', $admin);
     }
 
     public function update(AdminRequest $request, Admin $admin){
@@ -45,6 +49,8 @@ class AdminController extends Controller
     }
 
     public function destroy(Admin $admin){
+        if($admin->id == 1)
+            return back()->with('error', 'Você não pode excluir o super admin');
         $admin->delete();
         return back()->with('success', 'Registro excluído com sucesso');
     }
